refactor(web): 패키지와 클래스 이름을 servlet에서 controller로 변경
