+++ src/main/java/next/controller/ListUserController.java
@@ -1,6 +1,7 @@
-package next.web;
+package next.controller;
 
 import core.db.Database;
+
 import java.io.IOException;
 import javax.servlet.RequestDispatcher;
 import javax.servlet.ServletException;
@@ -8,17 +9,15 @@ import javax.servlet.annotation.WebServlet;
 import javax.servlet.http.HttpServlet;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
 
 @WebServlet(value = "/user/list")
-public class ListUserServlet extends HttpServlet {
+public class ListUserController extends HttpServlet {
 
   @Override
   protected void doGet(HttpServletRequest req, HttpServletResponse resp)
-      throws ServletException, IOException {
+    throws ServletException, IOException {
     req.setAttribute("users", Database.findAll());
-    RequestDispatcher requestDispatcher = req.getRequestDispatcher("/user/list.jsp");
-    requestDispatcher.forward(req, resp);
+    RequestDispatcher rd = req.getRequestDispatcher("/user/list.jsp");
+    rd.forward(req, resp);
   }
 }
